requirements.txt
absl-py==1.3.0
asgiref==3.5.2
asttokens==2.1.0
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.3
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
fonttools==4.38.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
grpcio==1.50.0
idna==3.4
imageio==2.22.4
ipykernel==6.17.1
ipython==8.6.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.0
jupyter_client==7.4.4
jupyter_core==4.11.2
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
nbclient==0.7.0
nbconvert==7.2.3
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.8.8
nltk==3.7
numpy==1.23.4
oauthlib==3.2.2
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
pathlib==1.0.1
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.32
protobuf==3.20.3
psutil==5.9.3
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.6
PyWavelets==1.4.1
pywin32==304
PyYAML==6.0
pyzmq==24.0.1
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scipy==1.9.3
seaborn==0.12.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.6.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
textblob==0.17.1
thop==0.1.1.post2209072238
tifffile==2022.10.10
tinycss2==1.2.1
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typing_extensions==4.4.0
tzdata==2022.6
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
